+++ library/gradle-plugin/src/main/java/io/mehow/laboratory/gradle/FeatureFactoryInput.kt
@@ -5,8 +5,6 @@ import io.mehow.laboratory.generator.FeatureFactoryModel
 import io.mehow.laboratory.generator.FeatureFlagModel
 import io.mehow.laboratory.generator.Visibility.Internal
 import io.mehow.laboratory.generator.Visibility.Public
-import org.gradle.api.Project
-import java.util.function.Predicate
 
 /**
  * Representation of a generated feature factory class.
@@ -23,17 +21,6 @@ public class FeatureFactoryInput internal constructor(
    * Sets package name of the generated feature factory. Overwrites any previously set values.
    */
   public var packageName: String? = null
-
-  internal var projectFilter = Predicate<Project> { true }
-    private set
-
-  /**
-   * Sets which Gradle projects should be included to contributing
-   * their feature flags to the generated feature factory.
-   */
-  public fun projectFilter(filter: Predicate<Project>) {
-    projectFilter = filter
-  }
 
   internal fun toBuilder(features: List<FeatureFlagModel>, simpleName: String) = FeatureFactoryModel.Builder(
       visibility = if (isPublic) Public else Internal,